Homework3upd

+++ src/main/java/com/edu/ulab/app/web/UserController.java
@@ -1,8 +1,13 @@
 package com.edu.ulab.app.web;
 
 import com.edu.ulab.app.facade.UserDataFacade;
+import com.edu.ulab.app.validators.BookRequestValidator;
+import com.edu.ulab.app.validators.UserBookRequestValidator;
+import com.edu.ulab.app.validators.update.UserBookUpdateRequestValidator;
 import com.edu.ulab.app.web.constant.WebConstant;
+import com.edu.ulab.app.web.request.BookRequest;
 import com.edu.ulab.app.web.request.UserBookRequest;
+import com.edu.ulab.app.web.request.update.UserBookUpdateRequest;
 import com.edu.ulab.app.web.response.UserBookResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -10,9 +15,14 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.MediaType;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import javax.management.openmbean.InvalidKeyException;
+import javax.validation.*;
 import javax.validation.constraints.Pattern;
+
+import java.util.List;
 
 import static com.edu.ulab.app.web.constant.WebConstant.REQUEST_ID_PATTERN;
 import static com.edu.ulab.app.web.constant.WebConstant.RQID;
@@ -23,9 +33,15 @@ import static com.edu.ulab.app.web.constant.WebConstant.RQID;
         produces = MediaType.APPLICATION_JSON_VALUE)
 public class UserController {
     private final UserDataFacade userDataFacade;
+    private final UserBookRequestValidator userBookRequestValidator;
+    private final UserBookUpdateRequestValidator userBookUpdateRequestValidator;
+    private final BookRequestValidator bookRequestValidator;
 
-    public UserController(UserDataFacade userDataFacade) {
+    public UserController(UserDataFacade userDataFacade, UserBookRequestValidator userBookRequestValidator, UserBookUpdateRequestValidator userBookUpdateRequestValidator, BookRequestValidator bookRequestValidator) {
         this.userDataFacade = userDataFacade;
+        this.userBookRequestValidator = userBookRequestValidator;
+        this.userBookUpdateRequestValidator = userBookUpdateRequestValidator;
+        this.bookRequestValidator = bookRequestValidator;
     }
 
     @PostMapping(value = "/create")
@@ -34,29 +50,67 @@ public class UserController {
                     @ApiResponse(description = "User book",
                             content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
                                     schema = @Schema(implementation = UserBookResponse.class)))})
-    public UserBookResponse createUserWithBooks(@RequestBody UserBookRequest request,
+    public UserBookResponse createUserWithBooks(@Valid @RequestBody UserBookRequest request,BindingResult bindingResult,
                                                 @RequestHeader(RQID) @Pattern(regexp = REQUEST_ID_PATTERN) final String requestId) {
+        userBookRequestValidator.validate(request,bindingResult);
         UserBookResponse response = userDataFacade.createUserWithBooks(request);
         log.info("Response with created user and his books: {}", response);
         return response;
     }
 
     @PutMapping(value = "/update")
-    public UserBookResponse updateUserWithBooks(@RequestBody UserBookRequest request) {
+    @Operation(summary = "Update user book row.",
+            responses = {
+                    @ApiResponse(description = "User book",
+                            content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
+                                    schema = @Schema(implementation = UserBookResponse.class)))})
+    public UserBookResponse updateUserWithBooks(@Valid @RequestBody UserBookUpdateRequest request,BindingResult bindingResult) {
+        userBookUpdateRequestValidator.validate(request,bindingResult);
+
         UserBookResponse response = userDataFacade.updateUserWithBooks(request);
         log.info("Response with updated user and his books: {}", response);
         return response;
     }
 
+    @PutMapping(value = "/update/UserBooks/{userId}")
+    public UserBookResponse updateUserBooks(@Valid @RequestBody List<BookRequest> request, BindingResult bindingResult, Long userId) {
+        if(userId<1){
+            log.error("Invalid index format");
+            throw new InvalidKeyException("Incorrect id");
+        }
+        request.forEach(bookRequest ->  bookRequestValidator.validate(bookRequest,bindingResult));
+
+        UserBookResponse response = userDataFacade.updateUserBooks(request,userId);
+        log.info("Response with updated user and his books: {}", response);
+        return response;
+    }
+
     @GetMapping(value = "/get/{userId}")
-    public UserBookResponse updateUserWithBooks(@PathVariable Long userId) {
+    @Operation(summary = "Get user book row.",
+            responses = {
+                    @ApiResponse(description = "User book",
+                            content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
+                                    schema = @Schema(implementation = UserBookResponse.class)))})
+    public UserBookResponse getUserWithBooks(@PathVariable Long userId) {
+        if(userId<1){
+            log.error("Invalid index format");
+            throw new InvalidKeyException("Incorrect id");
+        }
         UserBookResponse response = userDataFacade.getUserWithBooks(userId);
         log.info("Response with user and his books: {}", response);
         return response;
     }
 
     @DeleteMapping(value = "/delete/{userId}")
+    @Operation(summary = "Delete user book row.",
+            responses = {
+                    @ApiResponse(description = "User book",
+                            content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE))})
     public void deleteUserWithBooks(@PathVariable Long userId) {
+        if(userId<1){
+            log.error("Invalid index format");
+            throw new InvalidKeyException("Incorrect id");
+        }
         log.info("Delete user and his books:  userId {}", userId);
         userDataFacade.deleteUserWithBooks(userId);
     }
